chore(common): handles loadout trait cleanup
retire old `HandlesLoadout` in favor of new `HandlesLoadout` (previous `HandlesLoadout2`)
