+++ src/plugins/common/src/traits/accessors/get.rs
@@ -1,4 +1,5 @@
 mod assets;
+mod components;
 mod entity;
 mod handle;
 mod option;
@@ -6,10 +7,7 @@ mod ray;
 mod result;
 
 use bevy::{
-	ecs::{
-		component::Mutable,
-		system::{StaticSystemParam, SystemParam, SystemParamItem},
-	},
+	ecs::system::{StaticSystemParam, SystemParam, SystemParamItem},
 	prelude::*,
 };
 use std::ops::Deref;
@@ -122,15 +120,6 @@ pub trait ContextChanged {
 	fn context_changed(&self) -> bool;
 }
 
-impl<T> ContextChanged for Ref<'_, T>
-where
-	T: Component,
-{
-	fn context_changed(&self) -> bool {
-		self.is_changed()
-	}
-}
-
 /// Retrieve a context for data inspection.
 ///
 /// It is up to the implementor, what kind of system parameters are involved.
@@ -142,21 +131,6 @@ pub trait EntityContext<TMarker>: SystemParam {
 		entity: Entity,
 		marker: TMarker,
 	) -> Option<Self::TContext<'ctx>>;
-}
-
-impl<T, TMarker> EntityContext<TMarker> for Query<'_, '_, Ref<'static, T>>
-where
-	T: Component,
-{
-	type TContext<'ctx> = Ref<'ctx, T>;
-
-	fn get_entity_context<'ctx>(
-		query: &'ctx Query<Ref<T>>,
-		entity: Entity,
-		_: TMarker,
-	) -> Option<Self::TContext<'ctx>> {
-		query.get(entity).ok()
-	}
 }
 
 pub trait GetMut<TKey> {
@@ -178,36 +152,6 @@ pub trait EntityContextMut<TMarker>: SystemParam {
 		entity: Entity,
 		marker: TMarker,
 	) -> Option<Self::TContext<'ctx>>;
-}
-
-impl<T, TMarker> EntityContextMut<TMarker> for Query<'_, '_, Mut<'static, T>>
-where
-	T: Component<Mutability = Mutable>,
-{
-	type TContext<'ctx> = Mut<'ctx, T>;
-
-	fn get_entity_context_mut<'ctx>(
-		query: &'ctx mut Query<Mut<T>>,
-		entity: Entity,
-		_: TMarker,
-	) -> Option<Self::TContext<'ctx>> {
-		query.get_mut(entity).ok()
-	}
-}
-
-impl<T, TMarker> EntityContextMut<TMarker> for Query<'_, '_, &'static mut T>
-where
-	T: Component<Mutability = Mutable>,
-{
-	type TContext<'ctx> = Mut<'ctx, T>;
-
-	fn get_entity_context_mut<'ctx>(
-		query: &'ctx mut Query<&mut T>,
-		entity: Entity,
-		_: TMarker,
-	) -> Option<Self::TContext<'ctx>> {
-		query.get_mut(entity).ok()
-	}
 }
 
 pub trait TryApplyOn<'a, TKey>: GetMut<TKey> + 'a {
