--- src/plugins/skills/src/resources.rs
@@ -1,2 +1,0 @@
-pub(crate) mod skill_item_assets;
-pub(crate) mod skill_item_assets_usage;
