--- src/plugins/skills/src/resources/skill_item_assets.rs
@@ -1,8 +1,0 @@
-use crate::{item::Item, skills::Skill};
-use bevy::{ecs::system::SystemParam, prelude::*};
-
-#[derive(SystemParam)]
-pub struct SkillItemAssets<'w> {
-	pub(crate) items: Res<'w, Assets<Item>>,
-	pub(crate) skills: Res<'w, Assets<Skill>>,
-}
